apm-2134: to support increased rate limits implement caching on the client credentials jwks, and move pyproject to top level
