--- api_tests/docker-compose.yml
@@ -1,9 +1,0 @@
-version: "3"
-
-services:
-  test_env:
-    image: identity-service-api/test-env:latest
-    build: .
-    container_name: test_env
-    env_file:
-        docker.env
