--- api_tests/docker.env
@@ -1,5 +1,0 @@
-OAUTH_BASE_URI=
-OAUTH_PROXY=
-ID_TOKEN_PRIVATE_KEY_ABSOLUTE_PATH=
-JWT_PRIVATE_KEY_ABSOLUTE_PATH=
-APIGEE_API_TOKEN=
