--- api_tests/pyproject.toml
@@ -1,19 +1,0 @@
-[tool.poetry]
-name = "api_tests"
-version = "0.1.0"
-description = ""
-authors = ["Laurence Smith <laurence.smith@airelogic.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-requests = "^2.25.0"
-pytest = "^6.1.2"
-pytest-xdist = "^2.1.0"
-pytest-sugar = "^0.9.4"
-api-test-utils = {url = "https://github.com/NHSDigital/apim-test-utils/releases/download/v1.0.86-alpha/api_test_utils-1.0.86a0-py3-none-any.whl"}
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
