+++ azure/azure-pr-pipeline.yml
@@ -46,7 +46,7 @@ extends:
         post_deploy:
           - template: ./templates/run-integration-tests.yml
             parameters:
-              service_name: ${{ variables.service_name }}
+              test_type: e2e
       - environment: internal-dev
         service_name: ${{ variables.service_name }}-int
         short_service_name: ${{ variables.short_service_name }}-int
