+++ e2e/performance/pyproject.toml
