+++ e2e/scripts/response_bank.py
@@ -1,6 +1,6 @@
 """ A repository bank which hold large bits of information in a dictionary format.
 the keys are based on the name of tests to better identify what data belongs to which test."""
-from api_tests.scripts import config
+from e2e.scripts import config
 
 BANK = {
     'test_authorize_endpoint': {
