+++ e2e/tests/conftest.py
@@ -3,7 +3,7 @@ import asyncio
 from api_test_utils.oauth_helper import OauthHelper
 from api_test_utils.apigee_api_apps import ApigeeApiDeveloperApps
 from api_test_utils.apigee_api_products import ApigeeApiProducts
-from api_tests.scripts.generic_request import GenericRequest
+from e2e.scripts.generic_request import GenericRequest
 
 
 @pytest.fixture()
