+++ e2e/tests/oauth/test_jwt_unattended_access.py
@@ -1,4 +1,4 @@
-from api_tests.scripts.config import OAUTH_URL, ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH
+from e2e.scripts.config import OAUTH_URL, ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH
 import pytest
 from uuid import uuid4
 from time import time
