+++ package.json
@@ -15,11 +15,11 @@
     "speccy": "^0.11.0"
   },
   "devDependencies": {
-    "newman": "^4.5.7",
-    "newman-reporter-html": "^1.0.5",
-    "newman-reporter-htmlextra": "^1.11.3",
     "apigeetool": "^0.14.2",
     "license-checker": "^25.0.1",
-    "minimist": "^1.2.2"
+    "minimist": "^1.2.2",
+    "newman": "^4.5.7",
+    "newman-reporter-html": "^1.0.5",
+    "newman-reporter-htmlextra": "^1.11.3"
   }
 }
