+++ proxies/live/apiproxy/policies/ServiceCallout.ClientCredentialsGrantGetJWKS.xml
@@ -8,7 +8,7 @@
     </Set>
     <IgnoreUnresolvedVariables>false</IgnoreUnresolvedVariables>
   </Request>
-  <Response>calloutResponse</Response>
+  <Response>ccjwks_callout_response</Response>
   <HTTPTargetConnection>
     <SSLInfo>
       <Enabled>true</Enabled>
