+++ proxies/live/apiproxy/policies/VerifyJWT.ClientCredentials.xml
@@ -2,7 +2,7 @@
   <Algorithm>RS512</Algorithm>
   <Source>request.formparam.client_assertion</Source>
   <PublicKey>
-    <JWKS ref="calloutResponse.content" />
+    <JWKS ref="ccjwks" />
   </PublicKey>
   <Audience ref="proxy.url" />
   <TimeAllowance>10s</TimeAllowance>
