+++ proxies/live/apiproxy/proxies/default.xml
@@ -325,7 +325,19 @@
                     <Name>Javascript.SetServiceCalloutUrl</Name>
                 </Step>
                 <Step>
-                    <Name>ServiceCallout.ClientCredentialsGrantGetJWKS</Name>
+                  <Name>LookupCache.CCJWKS</Name>
+                </Step>
+                <Step>
+                  <Name>ServiceCallout.ClientCredentialsGrantGetJWKS</Name>
+                  <Condition>ccjwks = null</Condition>
+                </Step>
+                <Step>
+                  <Name>PopulateCache.CCJWKS</Name>
+                  <Condition>ccjwks = null</Condition>
+                </Step>
+                <Step>
+                  <Name>AssignMessage.CCJWKS</Name>
+                  <Condition>ccjwks = null</Condition>
                 </Step>
                 <!-- Not sure this is necessary
           <Step>
@@ -436,7 +448,19 @@
                 <Name>Javascript.SetServiceCalloutUrl</Name>
             </Step>
             <Step>
-                <Name>ServiceCallout.ClientCredentialsGrantGetJWKS</Name>
+              <Name>LookupCache.CCJWKS</Name>
+            </Step>
+            <Step>
+              <Name>ServiceCallout.ClientCredentialsGrantGetJWKS</Name>
+              <Condition>ccjwks = null</Condition>
+            </Step>
+            <Step>
+              <Name>PopulateCache.CCJWKS</Name>
+              <Condition>ccjwks = null</Condition>
+            </Step>
+            <Step>
+              <Name>AssignMessage.CCJWKS</Name>
+              <Condition>ccjwks = null</Condition>
             </Step>
             <Step>
                 <Name>VerifyJWT.ClientCredentials</Name>
@@ -464,19 +488,19 @@
             </Step>
             <Step>
                 <Name>ServiceCallout.cis2GetJWKS</Name>
-                <Condition>jwks = null Or jwksUpdated = false and idTokenIssuer  Equals "nhsCis2"</Condition>
+                <Condition>jwks = null Or jwks_kid_found = false and idTokenIssuer Equals "nhsCis2"</Condition>
             </Step>
             <Step>
                 <Name>ServiceCallout.nhsLoginGetJWKS</Name>
-                <Condition>jwks = null Or jwksUpdated = false and idTokenIssuer  Equals "nhsLogin"</Condition>
+                <Condition>jwks = null Or jwks_kid_found = false and idTokenIssuer Equals "nhsLogin"</Condition>
             </Step>
             <Step>
                 <Name>PopulateCache.TokenExchangeJWKS</Name>
-                <Condition>jwks = null Or jwksUpdated = false</Condition>
+                <Condition>jwks = null Or jwks_kid_found = false</Condition>
             </Step>
             <Step>
                 <Name>AssignMessage.JWKS</Name>
-                <Condition>jwks = null Or jwksUpdated = false</Condition>
+                <Condition>jwks = null Or jwks_kid_found = false</Condition>
             </Step>
             <Step>
                 <Name>VerifyJWT.SubjectToken</Name>
@@ -809,15 +833,15 @@
                 </Step>
                 <Step>
                     <Name>ServiceCallout.GetJWKS</Name>
-                    <Condition>jwks = null Or jwksUpdated = false</Condition>
+                    <Condition>jwks = null Or jwks_kid_found = false</Condition>
                 </Step>
                 <Step>
                     <Name>PopulateCache.JWKS</Name>
-                    <Condition>jwks = null Or jwksUpdated = false</Condition>
+                    <Condition>jwks = null Or jwks_kid_found = false</Condition>
                 </Step>
                 <Step>
                     <Name>AssignMessage.JWKS</Name>
-                    <Condition>jwks = null Or jwksUpdated = false</Condition>
+                    <Condition>jwks = null Or jwks_kid_found = false</Condition>
                 </Step>
                 <Step>
                     <Name>VerifyJWT.FromExternalIdToken</Name>
