+++ proxies/live/apiproxy/resources/jsc/validate-claim.js
@@ -2,12 +2,12 @@ jwksString = context.getVariable('jwks');
 jwksObj = JSON.parse(jwksString);
 kid = context.getVariable('jwt.DecodeJWT.FromSubjectTokenFormParam.decoded.header.kid');
 
-jwksUpdated = false;
+jwks_kid_found = false;
 
 for (var i = 0; i < jwksObj.keys.length; i++) {
   if (kid == jwksObj.keys[i].kid) {
-      jwksUpdated = true;
+      jwks_kid_found = true;
   }
 }
 
-context.setVariable('jwksUpdated', jwksUpdated);
+context.setVariable('jwks_kid_found', jwks_kid_found);
